refactor: Duration - phase 1 - migrated util methods
